pyproject.toml
[tool.poetry]
name = "shadow-api-mapper"
version = "1.0.0"
description = "Automated Shadow and Zombie API discovery tool using hybrid static/dynamic analysis"
authors = ["VILLEN Security <security@villen.dev>"]
license = "MIT"
readme = "README.md"
packages = [{include = "shadow_mapper", from = "src"}]
keywords = ["security", "api", "shadow-api", "zombie-api", "ast", "sast", "dast"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
]

[tool.poetry.scripts]
shadow-mapper = "shadow_mapper.cli:app"

[tool.poetry.dependencies]
python = "^3.11"
# CLI
typer = {extras = ["all"], version = "^0.11.0"}
click = ">=8.0.0,<8.2.0"  # Pin for compatibility with typer 0.11.x
rich = "^13.7.0"
# Data Models
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
# Networking
httpx = {extras = ["http2"], version = "^0.27.0"}
# Browser Automation
playwright = "^1.42.0"
# Parsing
tree-sitter = "^0.21.0"
tree-sitter-javascript = "^0.21.0"
tree-sitter-python = "^0.21.0"
tree-sitter-typescript = "^0.21.0"
# YAML/JSON handling
pyyaml = "^6.0.1"
# Async support
anyio = "^4.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
mypy = "^1.8.0"
ruff = "^0.3.0"
respx = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = ["E501", "B008", "B905"]

[tool.ruff.isort]
known-first-party = ["shadow_mapper"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
